requirements.txt
fastapi
uvicorn
pydantic
sqlalchemy
psycopg2-binary
fastapi_sqlalchemy
python-dotenv
python-multipart
requests